LayoutTests/storage/indexeddb/value-undefined-expected.txt

CONSOLE MESSAGE: The setVersion() method is non-standard and will be removed. Use the "upgradeneeded" event instead.
Test IndexedDB undefined as record value

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open(name)
db = event.target.result
request = db.setVersion('1')
Deleted all object stores.
objectStore = db.createObjectStore('foo');
result = objectStore.add(undefined, Infinity);
result = objectStore.get(Infinity);
value = event.target.result;
PASS value is undefined
result = objectStore.openCursor();
cursor = event.target.result;
PASS cursor.key is Infinity
PASS cursor.value is undefined
PASS successfullyParsed is true

TEST COMPLETE
